Cleanup Optimization dir
split out optimization classes into their own files
